Removing XDUser::getAllRoles
- Removed the `XDUser::getAllRoles` function and replaced it with
`XDUser::getAcls(true)`. Where the code specified that the public role should be
included in the results, `ROLE_ID_PUBLIC` was added to the results of
`getAcls(true)`.

- Updated the places that the results of `getAllRoles`
are utilized to account for the no longer having a `Role` object. This was
usually just removing `->getIdentifier()` as `getAcls(true)` returns the
equivalent of an array of $role->getIdentifier().

- Removed tests for the `getAllRoles` function as it's not longer present.