Updated Center[Director|Staff]Role getActiveCenter
These changes re-instate the behavior that was seen prior to updating the SQL
for these functions ( i.e. the only time an organization_id will be returned is
if the user in question is actually assigned the Center[Director|Staff] acl.
Else, -1 is returned ).

The test artifact that these functions use has been reverted accordingly.