Update logrotate to include a maxsize
These are preliminary proposed changes in response to XDMoD-2020-0003 and may be
subject to further change. While it still remains to be seen if we keep the
`/var/log/xdmod/apache*.log` config block / logs around it may be useful to
include this directive for the logs that XDMoD is directly creating / managing.