Taking care of a PHP Warning that was continually thrown
This was an odd one as it was a `continue` in a `case` statement located in a
`for` loop. A `break` would have sufficed to end the `case` statement. If we
take into account the warning message above `Skipping filter...` and that the
`for` loop is looping through filters we can, with reasonable certainty assume
that the intent was to `continue` the `for` loop. To make that happen we needed
to add the `2`.