Fixing SQL Errors
- */*-aggregation*.json: The sql statements constructed from this configuration
files were throwing a division by zero error. To resolve this problem an `IF`
statement was added that ensured the `task.wallduration` value was always at
least 1.

- post_ingest_update.sql: This statement wasn't operating as intended. After
some investigation I found that the `INTERVAL 1 SECOND` was not supported and
after discussion w/ Greg and some testing, changed it to `- 1` as the column
values are already in seconds.