Ensuring we test the contents of the log table
We ran into a problem where the records inserted into `log_table` were not
correctly formatted. This had the nock-on effect of making the log viewer tab in
the internal dashboard display useless as it could not parse the malformed
records.

I've modified this test so that the logger is retrieved in the same way as
loggers in the rest of the app and also added tests to ensure that the records
generated by said logger are in the format we expect.