Updates to handle the new authentication system
The previous function `CCR.invokePost` would do a pre-flight check to ensure
that a user was authenticated / authorized to make the call before actually
making it. That file is no longer needed / exists so this function call was
changed to be the same as what occurs when a user was determined to be
authorized. The authentication / authorization is now handled on the backend for
each rest call.