Updates to expected content-type / http code
The current rest stack defaults to setting the content-type to application/json
when it returns json so where before we would be getting back "json" but
expected the content-type to be `text/html` we can now update it to reflect the
new content-type.

Included in this commit are also some massaging of the expected http code, most
often it's expecting a 4** error code where before we had a 200.

There were also some places where we expected the returned value to be
stringified json that needed to be `json_decode`d, but now that it's actually
returning json it's decoded for us and we're presented with an array as opposed
to a string.

Some additional debugging was also added.