Adding Pre-Req changes for etl_viewer
ETLv2 wasn't really designed with introspection in mind, to help surface
the information contained within these classes I've implemented
`\JsonSerializable` for a portion of the ETLv2 class base. This will
then be used in a forthcoming `ETLControllerProvider` class to build the
underlying data structures that will power the ETL Visualizations.