Migrating / Removing code from security.php
So this is kind of a half-step commit, ideally there should be one basic
way (process) to get / authenticate a user. To get to this point will require
some additional work. As a step in the right direction I've pulled out the
`detectUser` function ( and the funcitons it depends on ) from
`libraries/security.php` and moved them to `src/BaseController.php` so that
authentication related code is in one place ( that place being the "new" Symfony
code ) and the Symfony controllers that still used `detectUser` have been
updated to use new `BaseController::detectUser` function.

A future commit ( or maybe PR ) will contain the migration of this code to be
more Symfony specific.