Updates to address user logged in related problems
These changes should address the following asana tasks:
-
https://app.asana.com/1/14740318781074/project/1211850781356424/task/1211881442164997?focus=true
-
https://app.asana.com/1/14740318781074/project/1211850781356424/task/1211881442164990?focus=true

The problems that were encountered were that certain css files were not present
when viewing the page before logging in. The ultimate cause of the problem was
that the logic used to determine whether or not a user was logged in had
changed. Previously it was based on the presence of the `xdUser` session
variable, in the new code it was based on if an XDUser was returned from the
`getXDUser` function ( i.e. was a null returned ), which was always the case.

The fix was to add a new template variable `user_logged_in` and set it based on
if there was an `xdUser` session variable ( and if the user is not public, since
the public user can't be logged in ) and then use this new boolean variable
where we had previously tested for `user is not null`.

After the above fix went in there was a problem w/ SSO Login. The observed
behevaior was that SSO Login would be successful, but the sign in link would not
change to the logged in users name.

The fix for this was to make sure that the `xdUser` session variable is set in
`BaseController.php::getXDUser`.